+++ Display-ILI9488-320x480.h
@@ -13,7 +13,7 @@ public:
   DisplayILI9488(uint16_t width, uint16_t height, uint8_t rotation, bool invertDisplayColors);
   void clearScreen(uint8_t color) override;
 
-  void showConnectScreen(uint32_t serialBaudRate, float currentVersion) override;
+  void showConnectScreen(uint16_t serialBaudRate, float currentVersion) override;
   void hideConnectScreen(void) override;
   void refreshConnectScreen(void) override;
